Rewrote source code for post route in listings to generate slug for each product, for seo
