+++ swap-yard-monolith/package.json
@@ -21,6 +21,7 @@
     "jose": "^6.1.3",
     "jsonwebtoken": "^9.0.3",
     "lucide-react": "^0.563.0",
+    "nanoid": "^5.1.7",
     "next": "16.1.6",
     "react": "19.2.3",
     "react-dom": "19.2.3",
